debugging
Fixed an extremely annoying bug. Turns out the official Arduino Keypad
library treats a key of 0 as a NO_KEY! Blarg I am dead.

Also changed the saveload toggle button into a standalone toggle switch,
instead of confusingly incorporating it into the butkey matrix.

Also reduced the number of preset buttons in butkeypad to 9, of which
I'll be using 8.

Finally, spiffed up the PWM action on the save/load LEDs.